add dynamic version to cli
This forcibly bumps the platform requirement to macOS 12. I'm fine with
that, since after a lot of consideration, it's unsupported by Apple
themselves, so I have no qualm in keeping Hedge (and, Quay as of right
now) near-ish currently supported platforms.

This could change in the future, like if I can split Quay and Hedge's
supported platforms, but it works for now.